+++ src/Campaign.php
@@ -7,8 +7,11 @@ namespace putyourlightson\campaign;
 
 use Craft;
 use craft\base\Plugin;
+use craft\console\Controller as ConsoleController;
+use craft\console\controllers\ResaveController;
 use craft\controllers\PreviewController;
 use craft\elements\User;
+use craft\events\DefineConsoleActionsEvent;
 use craft\events\DefineFieldLayoutFieldsEvent;
 use craft\events\FieldEvent;
 use craft\events\PluginEvent;
@@ -221,6 +224,10 @@ class Campaign extends Plugin
         // Register tracker controller shorthand for site requests
         if (Craft::$app->getRequest()->getIsSiteRequest()) {
             $this->controllerMap = ['t' => TrackerController::class];
+        }
+
+        if (Craft::$app->getRequest()->getIsConsoleRequest()) {
+            $this->registerResaveCommands();
         }
 
         if (Craft::$app->getRequest()->getIsCpRequest()) {
@@ -875,6 +882,62 @@ class Campaign extends Plugin
                 $event->permissions[] = [
                     'heading' => $this->name,
                     'permissions' => $permissions,
+                ];
+            }
+        );
+    }
+
+    /**
+     * Registers resave commands.
+     *
+     * @since 2.16.0
+     */
+    private function registerResaveCommands(): void
+    {
+        Event::on(ResaveController::class, ConsoleController::EVENT_DEFINE_ACTIONS,
+            function(DefineConsoleActionsEvent $event) {
+                $event->actions['campaigns'] = [
+                    'action' => function(): int {
+                        /** @var ResaveController $controller */
+                        $controller = Craft::$app->controller;
+                        $criteria = [];
+                        if ($controller->type !== null) {
+                            $criteria['campaignType'] = explode(',', $controller->type);
+                        }
+                        return $controller->resaveElements(CampaignElement::class, $criteria);
+                    },
+                    'options' => ['type'],
+                    'helpSummary' => 'Re-saves Campaign campaigns.',
+                    'optionsHelp' => [
+                        'type' => 'The campaign type handle(s) of the campaigns to resave.',
+                    ],
+                ];
+
+                $event->actions['mailing-lists'] = [
+                    'action' => function(): int {
+                        /** @var ResaveController $controller */
+                        $controller = Craft::$app->controller;
+                        $criteria = [];
+                        if ($controller->type !== null) {
+                            $criteria['mailingListType'] = explode(',', $controller->type);
+                        }
+                        return $controller->resaveElements(MailingListElement::class, $criteria);
+                    },
+                    'options' => ['type'],
+                    'helpSummary' => 'Re-saves Campaign mailing lists.',
+                    'optionsHelp' => [
+                        'type' => 'The mailing lists type handle(s) of the mailing lists to resave.',
+                    ],
+                ];
+
+                $event->actions['contacts'] = [
+                    'action' => function(): int {
+                        /** @var ResaveController $controller */
+                        $controller = Craft::$app->controller;
+                        return $controller->resaveElements(ContactElement::class);
+                    },
+                    'options' => [],
+                    'helpSummary' => 'Re-saves Campaign contacts.',
                 ];
             }
         );